requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
moviepy==1.0.3
opencv-python-headless==4.10.0.84
Pillow==11.1.0
numpy==1.26.4
pydub==0.25.1
noisereduce==3.0.3
librosa==0.10.2
soundfile==0.12.1
requests==2.32.3
aiofiles==24.1.0
python-dotenv==1.0.1
jinja2==3.1.5
faster-whisper==1.1.0
